Delete README.md
This project uses four photo sensors to aim a solar panel toward the sun. It will "park" itself and turn off the servos when the sunlight level is low in order to conserve energy usage. It also triggers an LED to signal a raspberry pi weather station. The LED is on for high light level and off for mid or low light level.